Add files via upload
This is a Lakers-Themed Drawing App that is interactive graphical application development using Java Swing and 2D Graphics.The basket-ball related app enables users to choose from shapes (Basketball or Hoop) on a designed court. The colors that can be chosen for this project are either yellow or purple (Lakers colours). The user is allowed to choose what shape and what colour they would like to use. The canvas can also be cleared or reset to default settings using the keyboard shortcuts or by mouse clicking the buttons.

    * Keyboard shortcuts
   ‘b’ : Selects Basketball shape
   ‘h’ : Selects Hoop shape
   ‘p’ : Purple colour change
   'g’ : Gold colour change
   ‘c’ : Canvas Clears
   ‘r’ : Reset to default settings